[v14] [Docs] Update documentation for max duration feature in access requests (#31679)
* Update documentation for max duration feature in access requests

Documentation for the access request feature has been expanded to describe the 'max_duration' property in role requests. The change explains how to override the default behaviour (where a user can request access to a role until the session expires) by setting a 'max_duration'. Details include how to set this in the web UI or CLI, the need to define the setting in the role definition, and its interaction with the 'max_session_ttl'. The update provides necessary guidance for users requiring elevated access beyond their current session.

* Apply code review suggestions.