Remove bypass workflow for integration and docs tests (#31953)
* Remove bypass workflow for integration tests

It turns out, the 'paths' filter is not supported with the merge
queue. As a result, we're running a lot more actions in the merge
queue than we thought we were. For example, integration tests are
running even if the PR(s) in the queue only contain docs changes.

Instead of relying on GitHub to do the filtering (which isn't supported)
and using the "multiple workflows with the same name" hack, we use a
third party action to detect what files were changed and determine
whether subsequent jobs should run.

* Remove doc-tests bypass

This will allow us to test both cases

* Address review comments