Append a portion of the GUID when naming discovered desktops
Some users have encountered cases where they connect multiple
distinct AD domains to Teleport, but the name of the domain is
the same and the name of some of the hosts within the domain are
also the same.

As a result, Teleport incorrectly deduplicates these desktops in
the UI (we can't distinguish them from identical desktops reported
by multiple agents).

Fix this by appending a portion of the objectGUID to the desktop name.
Even if mutiple domains have identical names and hostnames, the GUID
will differ and both will be shown in the UI.