Prevent duplicate service registration (#32050)
Enterprise tests do not always set the enterprise build type which
causes test failures due to duplicate service registration. To
prevent every test from needing to update the modules the PluginRegistry
was modified to expose a mechanism to detect if a plugin is registered
or not. The old check that registered a not implemented LoginRuleService
now makes use of this mechanism to verify that the enterprise auth
plugin is not registered.