Remove unused images from the docs
Used the following command to find unused images. This lists all image
files in the docs and, for each image, searches all docs pages for
references to the image file (which include the `/img` path segment but
usually not `docs`). If there are no matches, it prints the image file
name:

```bash
$ find docs/img -name "*.png" | sed -E "s|docs/(.*)|\1|" | xargs -I{} bash -c '
grep -qR "{}" docs/pages; if [ $? -ne 0 ]; then echo {}; fi'
```

Also fixes a minor linter error that did not come up in earlier commits.