[v14] AWS OIDC: command to configure IAM for listing databases (#31981)
* AWS OIDC: command to configure IAM for listing databases

Previously we were guiding users in a series of screens and copy/paste
so that the user could configure the AWS OIDC Integration.
This guide also included adding the first feature: enroll an RDS
Database.

We created a new command that sets up the OIDC IdP, but decoupled from
the enroll rds feature.

After updating the guide to use the new command and remove the guide, we
must provide a way for users to set up the required policy for Enrolling
RDS Databases.
This PR adds that command

* review pt1

* rename to ConfigureIAMListDatabasesRequest

* removed wrapper

* fix command description

* move aws client setup to common

* use logrus instead of log