Upload one file per Vercel preview deployment (#32231)
Currently, the Vercel preview workflow uses the default `vercel deploy`
behavior, which uploads every source file to Vercel. The result is often
rate limiting by the Vercel API, which prevents some preview workflows
from running.

This change adds the `--archive=tgz` flag to the `vercel deploy` command
to upload a single tarball instead, as recommended by Vercel support.

The flag is undocumented, but you can consult the PR that added it to
the `vercel` CLI (vercel/vercel#8356) for context.