[v13] Use the instance role for the upload completer (#32346)
* Use the instance role for the upload completer

When we originally moved to running one uploader per process
the instance role wasn't fully supported, so we would pick
whatever client was available first.

Now that the instance role is supported in v11+, we can make
things simpler and more deterministic by always using the
instance role.

Note: the instance client isn't available when the auth server is
running in-process - so we use the auth server directly in those
cases. This is more performant anyway.

* Log a warning if the instance connector is not ready on time