Create utility function for calculating next Access List review date.
A utility function has been created which will select the next Access List
review date given the audit frequency and the current review date. The logic
is fairly complex, but the intention here is to make sure we always choose
N months after the current review date on the same day of the month, with
some assumptions being made.

The tests are fairly comprehensive here, testing a large swath of years in
durations that we don't expect to see in reality with numbers close to
overflowing `time.Duration`.