Add optimistic locking support to sqlite backend (#31984)
A new revision column is added to the kv table to accomodate
storing the resource revision. A schema migration for the table
was added to ensure any existing databases have the table added.
Additionaly the meta table and its associated import functionality
was removed since it was unused. Various other small clean up
operations were also done to eliminate warnings and remove unused
functionality.

Contributes to #30416.