Update access list guide.mdx
Users with older versions of the editor role are confused when they can see the access list button but it is greyed out. I amended the guidance to specify which version of the editor is preconfigured with access list permissions as well as added instructions for adding the permissions to older role versions.