[v14] Secure File Removal Improvements (#32435)
* utils.RemoveSecure: Still attempt a removal after error in overwrite

As extra caution, even if an error occurs during the overwrite process, we still want to attempt a removal of sensitive files.

* keystore.go: More secure removal of keyfiles

This commit ensures that deleted keyfiles have been overwritten.  This has little value on SSD's but can improve the security when the disk is magnetic.

* Apply PR feedback, notably better testing and early unlinking if possible

This adds an OS conditional so that if possible the file will be removed and then overwritten using the previous file handle.
This will reduces the chance that the file will be witnessed with unexpected contents.