Begin keyless signing and verification of images using GitHub OIDC tokens (#2210)
Have `cosign` use keyless signing and verification of images.
https://docs.sigstore.dev/signing/overview/

This also means that all signature events will be logged in Rekor, a signature
transparency log, providing an auditable record of when a signature was created.

Note that existing signing/verification methods using the signing key are left
intact. This just adds another additional layer. We may later choose to remove
the signing key entirely, but it is not in-scope at this time.

Source-commit: gravitational/teleport.e@e47569808d0036213e436585514ff5f558949a9e
Source-pull-requests: gravitational/teleport.e#2210