[v14] Remove non-file path links from partials (#32234)
Backports #31708

* Remove non-file path links from partials

Fixes #12618

Ensure that all partials use relative links to MDX files for more
accurate linting. Only two partials contain links to URL paths, and one
of these is not included anywhere in the docs.

This change removes the unused partial and changes the remaining link to
a file path.

* Fix linter error