Improve makefile to use all available processors
The commit modifies the make commands in the Github Action Workflow script to use all available processors for the build process. By using the "-j$(nproc)" option, the build process should be significantly quicker as it's now capable of parallel processing. This change gets the most out of available resources and improves build efficiency.