[v13] Update Okta SDK to v2.20.0
Fixes unexpected EOF responses seen when the Okta service attempts to interact
with the Okta API. This only affects v13, as v14 and master are both running
v2.20.0 already.

https://github.com/okta/okta-sdk-golang/issues/363 is an issue filed against
the Okta Go SDK where some API calls were seeing empty responses or unexpected
EOFs for larger responses. https://github.com/okta/okta-sdk-golang/pull/369 was
submitted to fix this, and then was pulled in as part of
https://github.com/okta/okta-sdk-golang/pull/372.