Deflake `TestInteractiveSessionsNoAuth` test
For this test, the LockWatcher is marked as stalled. When in stalled
mode, the watcher bypasses cache and hits auth server directly.
During the test, the auth rate limit is exceeded which causes the
watcher to fail and report unexpected errors.

This PR bumps the auth server limits to prevent these cases.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>