Prevent Kube proxy from set the default Kube impersonation headers (#32839)
This PR removes the ability of the Kubernetes Proxy to set the default
values of Kubernetes Impersonation headers. This task is delegated to
the kubernetes_service or kubernetes legacy proxy that handles the
request.

This is required because under certain contiditons it might have
requests that failed because the values were improperly defined.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>