Add hosted plugin instructions to the Web UI (#2116)
* Add hosted plugin instructions to the Web UI

See gravitational/teleport#29979

The Web UI view for enrolling hosted plugins currently links to the docs
for setup instructions. This is inappropriate for hosted plugin users,
since the instructions in the docs are intended for self-hosted plugins.

Further, adding hosted plugin instructions to the existing docs would
only add another branch to pages that already include several, e.g., for
Kubernetes vs. Linux server installations.

This change adds instructions to the Teleport Web UI for setting up
third-party tools in order to install hosted plugins. To do so, it edits
the view for enrolling plugins:

- Adds a `Setup` field to the `HostedPlugin` type so we can consistently
  include setup information apart from each plugin's description.
- Divides the "Enroll" view for each plugin into subheadings to make it
  easier to read.
- Adapts the setup instructions from each hosted access request plugin's
  docs page for the "Enroll a plugin" view.

Note that this change does not include new images in the Web UI. This is
because the images included in the setup instructions within the docs
are not intended for the Web UI, and we don't have a way to make these
images look good there. We also don't have a good way to host avatar
images for downloading from the Web UI, so this change does not include
instructions for downloading avatar images for the Mattermost and
Discord plugins.

* Add Jamf instructions to the Web UI

* Add Okta Next Steps to the Web UI

Link to the docs for adding an import rule. This is not ideal, but the
Web UI does not yet support the UX required to render instructions for
creating an import rule, e.g., support for code snippets.

* Change instructions to ordered lists

* Fix linter issues

* Remove UI component references from instructions

- Opsgenie
- Mattermost
- Jira
- PagerDuty
- Discord

Source-commit: gravitational/teleport.e@8db9e4224b9070e0714c0651ba36aff5521ca01d
Source-pull-requests: gravitational/teleport.e#2116