[v14] Connect My Computer: Improve copy and UI consistency (#32890)
* CMC Status: Use namespace import for icons

* CMC Status: Use vertical three dots icon for the menu

This makes it consistent with the vertical three dots icon we use for
additional actions in the top bar.

* Change the name of agent cleanup daemon logfile

"cleanup" alone is too vague.

* Pass --insecure to agent if Connect runs in insecure mode

* Update copy related to security

* Sprinkle some whitespace in connectMyComputer context

* Explain why we can't remove connections before killing the agent

* Mention file URL in download error message

Without this, the UI would just show the previous message, but the user
would have no idea what exactly was being downloaded. This was especially
the case when updating the agent after it was already set up.