[v13] Track user connections across clusters (#32996)
* Break client facing types out of reversetunnel package

Backports part of #28357 to branch/v13

* Track user connections across clusters (#32688)

Ensure that `proxy_ssh_sessions_total` is incremented in both the
root and the leaf cluster if a session to a node in the leaf cluster
is established through the root cluster. This now also prevents
leaf clusters from terminating during a graceful shutdown if there
is an active connection that was created from the root.

Fixes #32605.

* Track user connections across clusters (#32688)

Ensure that `proxy_ssh_sessions_total` is incremented in both the
root and the leaf cluster if a session to a node in the leaf cluster
is established through the root cluster. This now also prevents
leaf clusters from terminating during a graceful shutdown if there
is an active connection that was created from the root.

Fixes #32605.

* Track connections to direct dial nodes across clusters

#32688 added support for tracking connections to sessions via a
reverse tunnel node but failed to account for sessions to direct
dial nodes.