[v14] Reliability improvements for HSM tests (#33091)
* log message improvements

* fix etcd cleanup

* re-enable TestHSMDualAuthRotation

* retry client connection tests

* fixes based on code review

* make fix-imports

* fix: use EventuallyWithT

* set short polling period

* fix: typo in error check

While working on https://github.com/gravitational/teleport/pull/32911 I
noticed the following broken error comparison, which would always log
because the error was always wrapped and never exactly matched
`http.ErrServerClosed`:

```
if err != nil && err != http.ErrServerClosed {
    log.Warningf("TLS server exited with error: %v.", err)
}
```

I tried to fix it with `errors.Is`, but unfortunately got the condition
inverted. This fixes the condition to match the original intended
behaviour. It's not useful to log `http.ErrServerClosed` errors because
that error is returned every time the server is manually closed.