Up-revs the Okta plugin settings version
Adds a version field to the OktaPlugin settings and updates the associated
protocol files and tests. This is in preparation for adding new behaviour
to the Okta plugin, and will allow Teleport to determine if a plugin
installation was created by the current version of Teleport (and should
get the new behaviour), or an old version (which will get no surprising
behavioural changes)

changing the behaviour of the Okta plugin depending
if the plugin is created from a current or old version of Teleport.