Remove commented code in service.go
Cleaned up service.go by removing unused, commented-out code that pertains to the AccessGraphProcessor. This is part of a larger effort to clean up the codebase and improve readability. The code about reusing the connection and adding TLS are marked as TODOs, but they are not yet implemented or used, hence also have been removed. This also safeguards from potential bugs in the future due to confusion. Now the connection gets explicitly closed everytime.