[v14] utils.RecursiveChown: Fix for Privilege Escalation due to following symlinks (#33248)
* utils.RecursiveChown: Harden against user access race conditions

Prior to this change a user could exploit Teleports privileged access to `chown` arbitrary files on the system.

This is due to the directory being changed first, allowing a small time window where a user can remove or rename the still `root` owned files with a symlink.  The added tests help show this issue in a more controlled way.

A switch to `os.Lchown` avoids the risk in following symlinks to files.  In addition, in order to remove the risk for hardlinks (notably on OSX with reduced hardlink protections), as well as risks with directory symlinks, the folder structure is inspected before any `chown` operation.  And then the files are updated before their parent directories.

* Update other instances of `os.Chown` to `os.Lchown`

None of these cases should expect a symlink that would need to be followed.