Implement synchronization of access requests to the new AccessGraph service
Added a new function, `sendAccessRequests()`, to send all existing access requests to the AccessGraphService upon initialization. This is to ensure that the access requests are in sync on both the auth server and the new AccessGraph service. Also, `Namespace` parameter was added in the `listReq` in `getResourceEvents()` to specify the namespace of resources to retrieve, addressing missing details in the previous version. Removed `sendNodes()` function as resources should now be fetched via the new AccessGraph service.