Add GetFile RPC to AccessGraphService
This commit adds a new RPC `GetFile` to the `AccessGraphService`. This method allows clients to retrieve static UI files from the access graph container. A message was also added to facilitate requests for file retrievals (`GetFileRequest`) which takes a file path as a string, and a corresponding response message (`GetFileResponse`) returns the file's data in bytes. The changes were made in the respective proto, grpc and pb files.