ci: Add GHA workflows for linux tag builds (#2302)
Extend the workflows that do the linux push builds to do the tag builds
as well. The workflows are sufficient to replace the equivalent ones in
Drone and produce the same release artifacts. There are not separate
workflows for packaging deb/rpm archives as there are in Drone - that
is all taken care of in the single `build-linux.yaml` workflow.

A subsequent PR in the teleport repo will update `dronegen` to call the
`release-linux.yaml` workflow with appropriate parameters for each
build.

Drone duplicated a couple of builds - the amd64 and centos7 builds
ultimately called the same makefile targets for their builds. All amd64
builds are done on centos7 even if the workflow/pipeline was not named
with "centos7".

Issue: https://github.com/gravitational/teleport/issues/20729

Source-commit: gravitational/teleport.e@3badcee5b79a3288b24ee4580020466ce42735b0
Source-pull-requests: gravitational/teleport.e#2302