Refactor desktop audit event emission (#33189)
Our methods for emitting audit events take 10 arguments already,
and we need to add more as part of the work in #30417. To make
this more manageable, create an auditor struct that will hold on
to state that is shared for all audit events in a session (ID,
user identity, the desktop we're connecting to, etc.)

As a result, the "audit cache" for directory sharing events is
also simplified - we now create one of these per-session rather
than maintaining one large cache for all sessions.