Add SendEvent to AccessGraphService (#33193)
Updated several proto files and service.go to support the addition of the SendEvent method in AccessGraphService. This method enables sending an event to the access graph service. Also updated the access graph watcher to include the method and send events as necessary. These changes provide the mechanisms needed for proper interaction with the access graph service. Updated several proto files and service.go to support the addition of the SendEvent method in AccessGraphService. This method enables sending an event to the access graph service. Also updated the access graph watcher to include the method and send events as necessary. These changes provide the mechanisms needed for proper interaction with the access graph service.