ci: Fix determination of FIPS/Non-FIPS build (#2415)
Boolean outputs for steps in GitHub Actions are not actually booleans,
but strings. Job outputs are boolean, so when the FIPS checking logic
was moved from a job output to a step output, the logic for checking
FIPS was wrong - you need to compare against the strings `'true'` and
`'false'` explictly.

Source-commit: gravitational/teleport.e@4d74262424077370a5159272712143d7d1e05165
Source-pull-requests: gravitational/teleport.e#2415