ci: Ensure workflow tag is used as step name (#2421)
Put the workflow tagging step back into its own job, as this is rather
critical to identifying the workflow in `gh-trigger-workflow`. At the
very least, the name of the step must be the workflow tag, but to be
conservative, use a separte job like we do in other workflows. Without
this, we continually poll the actions API looking for the workflow.

Source-commit: gravitational/teleport.e@5bb5001c0f2fcecf6863d1d6d01c293d3a76b44e
Source-pull-requests: gravitational/teleport.e#2421