Gracefully handle web socket closure by clients
Updates the terminal stream to detect when the web socket is closed
by the UI via setting the close handler. This reduces the time it
takes the server to know the client closed the connection from the
value of the keep alive interval in the networking config (default 5m)
to as soon as a user closes the tab. Prior to this any session that
was terminated by closing the browser tab would have an active
session tracker until the web socket ping loop, which runs at the
keep alive interval, failed. Some logging was also cleaned up to
reduce the amount of spam that appears as a result of a session
being terminated by a user.