[v14] OpenTelemetry Updates (#33523)
* Bump go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp

Bumps [go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp](https://github.com/open-telemetry/opentelemetry-go-contrib) from 0.42.0 to 0.44.0.
- [Release notes](https://github.com/open-telemetry/opentelemetry-go-contrib/releases)
- [Changelog](https://github.com/open-telemetry/opentelemetry-go-contrib/blob/main/CHANGELOG.md)
- [Commits](https://github.com/open-telemetry/opentelemetry-go-contrib/compare/zpages/v0.42.0...zpages/v0.44.0)

---
updated-dependencies:
- dependency-name: go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>

* Bump go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp

Bumps [go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp](https://github.com/open-telemetry/opentelemetry-go-contrib) from 0.42.0 to 0.44.0.
- [Release notes](https://github.com/open-telemetry/opentelemetry-go-contrib/releases)
- [Changelog](https://github.com/open-telemetry/opentelemetry-go-contrib/blob/main/CHANGELOG.md)
- [Commits](https://github.com/open-telemetry/opentelemetry-go-contrib/compare/zpages/v0.42.0...zpages/v0.44.0)

---
updated-dependencies:
- dependency-name: go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>

* Additional OpenTelemetry updates

This updates other OpenTelemetry dependencies that Dependabot did not automatically open.  I believe it's safest if we generally just update all of OpenTelemetry to the latest.

* sync `api` OpenTelemetry versions to parent versions

* Bump go.opentelemetry.io/otel/sdk from 1.16.0 to 1.17.0 (#31185)

* AWS OIDC - List EC2: add instance id as label (#31415)

When listing EC2 instances we should be able to tell which ones were
already added.
To do so, we should use the instance-id which is unique.
We can't use the Spec.CloudMetadata.AWS.InstanceID because predicate
can't use fields that are optional.

This PR adds the InstanceID as label.
It uses `teleport.dev/instance-id` label key.

When creating a Node from WebUI, it will send the same set of labels
that was received.
So, WebUI will send this same label.

When listing EC2 instances, WebUI queries the backend for a list of
Nodes of type EC2 EICE and the same `instance-id` and uses the result to
show which ones were already added.

* update semconv version to match otel

---------

Co-authored-by: Marco André Dinis <marco.dinis@goteleport.com>
Co-authored-by: Tim Ross <tim.ross@goteleport.com>

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Co-authored-by: Marco André Dinis <marco.dinis@goteleport.com>
Co-authored-by: Tim Ross <tim.ross@goteleport.com>