Swap actions/create-github-app-token for tibdex/github-app-token (#32106)
tibdex is some random developer.  We prefer 2nd party actions from
GitHub, as we have a contractual relationship with them.

As part of this change, I'm also comfortable dropping the SHA pinning --
since the `actions` org can be held to a higher level of trust for
both security and backwards compatibility concerns.