Swap tibdex/github-app-token to actions/create-github-app-token (#2244)
* Swap actions/create-github-app-token for tibdex/github-app-token

tibdex is some random developer.  We prefer 2nd party actions from
GitHub, as we have a contractual relationship with them.

As part of this change, I'm also comfortable dropping the SHA pinning --
since the `actions` org can be held to a higher level of trust for
both security and backwards compatibility concern

* Change repository argument to new syntax

This feature was just added to create-github-app-token in
https://github.com/actions/create-github-app-token/pull/46, and it uses
slightly different syntax than tibdex's action.

Source-commit: gravitational/teleport.e@860b84b4f88108a49fa67104f4b88a1f6583220a
Source-pull-requests: gravitational/teleport.e#2244