Enable optimistic locking support for OIDC connectors (#33458)
Updates tctl edit and the web ui to start using optimistic locking.
The functionality to support optimistic locking already existed, the
APIs used by both clients were updated to use create/update instead
of upsert so that optimistic locking could be enforced. Most of the
changes introduced are tests to ensure that tctl edit, tctl create
behave as expected.

Note: the web ui changes are include in the e ref update.

Contributes to #30416.