Add an Access Request configuration guide (#33414)
* Add an Access Request configuration guide

Fixes #22496
Fixes #13927
Fixes #6557
Fixes #31095
Fixes #29980
Fixes #17630

The current approach to documenting Access Requests is to include how-to
guides for various scenarios. The downside of this approach is that we
don't have a great place to put general conceptual discussions of Access
Request configuration fields. This leads to confusion among users
regarding the way Teleport handles certain Access Request configuration
options.

This change adds a conceptual guide that explains all of the fields in a
Teleport role that are relevant to configuring Access Requests.

This change also moves conceptual discussions from other guides into the
new guide. To limit the scope of this change, it is not intended to
overhaul the existing guides. If a discussion of a particular
configuration field was buried in another guide, this change moves it
into the new guide:
- `preview_as_roles` discussion in the Role Requests guide
- TTL information in the Role Requests guide
- The reference role in the Role Requests page

* Respond to zmb3 feedback

- Use clearer wording when describing key concepts.
- Include more detail about the `max_duration` field.
- Use consistent version numbers in example roles.

* Fix linter issues