Fix table alignment in `tctl tokens ls` examples (#33654)
Fixes #32640

On two docs pages, `tctl tokens ls` examples use a built-time variable
to specify a preset token value. The pages space the values of the
example tables to anticipate the variable, not the variable's
replacement value, meaning that the tables appear incorrect on the
rendered docs site. This change fixes the spacing within the tables.