desktop access: soften the disconnect error
If the Windows host (RDP server) decides to terminate the connection,
it often sends a message indicating why the disconnect is coming.
One example of a server-initiated disconnect would be when the user
clicks "log off" in the RDP session or when they restart the remote
machine.

We forward this message to the UI, but we put it in a big red ERROR
box that makes it look like something went wrong.

Errors should be reserved for unexpected disconnections, so covert
the dialog to "info" level and soften the messaging.